pyproject.toml
[project]
name = "openai-stt-tts-test"
version = "0.1.0"
description = "Simple project for testing for STT/TTS"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "openai>=1.40.0",
    "python-dotenv>=1.2.1",
]
